Desenvolvimento da migracao de employees

+++ internal/repository/employees/delete.go
@@ -1,21 +1,22 @@
 package repository
 
-import "database/sql"
+import models "github.com/maxwelbm/alkemy-g6/internal/models"
 
 func (e *EmployeesRepository) Delete(id int) (err error) {
+	var exists bool
+	err = e.DB.QueryRow("SELECT EXISTS(SELECT 1 FROM employees WHERE id = ?)", id).Scan(&exists)
+	if err != nil {
+		return
+	}
+	if !exists {
+		err = models.ErrEmployeesNotFound
+		return
+	}
+
 	query := "DELETE FROM employees WHERE id = ?"
-	result, err := e.DB.Exec(query, id)
+	_, err = e.DB.Exec(query, id)
 	if err != nil {
 		return
-	}
-
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return
-	}
-
-	if rowsAffected == 0 {
-		return sql.ErrNoRows
 	}
 
 	return nil
